// src/app/core/models/IspitRad.ts
export class IspitRad {
    radId?: number;
    vrijemePostavljanjaRada?: string;
    urlRada?: string;
    ispitId?: number;
    naslov?: string;
    opis?: string;
    vrijemePocetka?: string;
    vrijemeZavrsetka?: string;
    urlIspita?: string;
    organizacijaKursa?: string;
    organizacijaKursaId?: number;
}